pyproject.toml
[project]
name = "rag-fast-flow"
version = "0.1.0"
description = "Streamlit app for comparing pure LLM vs RAG-enhanced responses for Fast Flow methodologies"
requires-python = ">=3.11"
dependencies = [
    "streamlit>=1.28.0",
    "qdrant-client>=1.7.0",
    "llama-index-core>=0.10.0",
    "llama-index-embeddings-ollama>=0.1.0",
    "requests>=2.31.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = []
